pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Sérgio Rafael Siqueira <osergiosiqueira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
redis = "^4.6.0"
pendulum = "^2.1.2"
dateparser = "^1.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
